Increase RelativePopulation precision from int to double
Increase precision to facilitate enabling SSM on stable at a less then 1% enable rate.

Bug: 338233130
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6049428
Commit-Queue: Mathijs Affourtit <maffourtit@google.com>
Cr-Commit-Position: refs/heads/main@{#1391829}

+++ chrome/common/profiler/thread_profiler_platform_configuration.h
@@ -36,14 +36,11 @@ class ThreadProfilerPlatformConfiguration {
   // - Within the enabled population, profiling is always enabled.
   // - Within the disabled population, profiling is always disabled.
   // - The experiment population is further split down to N equal-sized
-  // subgroups with different configurations. Often, N = 2, where one subgroup
-  // is disabled and one is enabled, but it could be more than 2 if we're
-  // experimenting with more specific details of how and where we enable the
-  // profiler. |experiment| must be divisible by N.
+  // subgroups with different configurations.
   struct RelativePopulations {
-    int disabled;
-    int enabled;
-    int experiment;
+    double disabled;
+    double enabled;
+    double experiment;
   };
 
   virtual ~ThreadProfilerPlatformConfiguration() = default;
